move test ... was in wrong folder
