+++ digicampipe/tests/resources/example_10evts.fits.fz
